Move BloctreeError into blocktree_db
Fewer dependency cycles
